requirements.txt
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.7.2
appdirs==1.4.4
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.22.0
attrs==20.3.0
black==20.8b1
cffi==1.14.5
chardet==4.0.0
chess==1.6.1
click==7.1.2
discord-components==0.5.3
discord-py-slash-command==2.3.2
discord.py==1.7.1
Flask==1.1.2
gunicorn==20.1.0
idna==3.1
isort==5.9.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mpmath==1.2.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycparser==2.20
PyNaCl==1.4.0
python-chess==1.999
python-dotenv==0.17.0
PyYAML==5.4.1
regex==2021.4.4
six==1.15.0
sympy==1.8
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.3
